refactor(app): split utils
