+++ app/src/scenes/admin-mission/index.jsx
@@ -14,7 +14,7 @@ import { STATUS_PLR } from "@/constants";
 import api from "@/services/api";
 import { captureError } from "@/services/error";
 import { compactMissionFilters, searchMissions } from "@/services/mission";
-import exportCSV from "@/services/utils";
+import exportCSV from "@/utils/csv";
 
 const TABLE_HEADER = [
   { title: "Mission", key: "title.keyword", width: "35%" },
