+++ app/src/scenes/auth/ForgotPassword.jsx
@@ -3,7 +3,7 @@ import { RiCheckboxCircleFill, RiErrorWarningFill } from "react-icons/ri";
 
 import api from "@/services/api";
 import { captureError } from "@/services/error";
-import { isValidEmail } from "@/services/utils";
+import { isValidEmail } from "@/utils/string";
 
 const Forgot = () => {
   const [values, setValues] = useState({ email: "" });
