+++ app/src/scenes/auth/Login.jsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { RiErrorWarningFill } from "react-icons/ri";
 import { Link, useNavigate } from "react-router-dom";
 
-import { toast } from "@/services/toast";
 import api from "@/services/api";
 import { captureError } from "@/services/error";
 import useStore from "@/services/store";
-import { isValidEmail } from "@/services/utils";
+import { toast } from "@/services/toast";
+import { isValidEmail } from "@/utils/string";
 
 const Login = () => {
   const [loading, setLoading] = useState(false);
