+++ app/src/scenes/user/Create.jsx
@@ -9,8 +9,8 @@ import LabelledSelect from "@/components/form/LabelledSelect";
 import Table from "@/components/Table";
 import api from "@/services/api";
 import { captureError } from "@/services/error";
-import { isValidEmail } from "@/services/utils";
 import { withLegacyPublishers } from "@/utils/publisher";
+import { isValidEmail } from "@/utils/string";
 
 const TABLE_HEADER = [{ title: "Nom" }, { title: "Roles", position: "center" }];
