+++ app/src/scenes/warnings/index.jsx
@@ -8,7 +8,7 @@ import { DAYS, MONTHS, WARNINGS, YEARS } from "@/constants";
 import api from "@/services/api";
 import { captureError } from "@/services/error";
 import useStore from "@/services/store";
-import { slugify } from "@/services/utils";
+import { slugify } from "@/utils/string";
 import { Link } from "react-router-dom";
 
 const LINKS = {
