+++ app/src/utils/date.js
@@ -1,0 +1,15 @@
+export const timeSince = (date) => {
+  let seconds = Math.floor((new Date() - date) / 1000);
+  let interval = seconds / 31536000;
+
+  if (interval > 1) return Math.floor(interval) + " années";
+  interval = seconds / 2592000;
+  if (interval > 1) return Math.floor(interval) + " mois";
+  interval = seconds / 86400;
+  if (interval > 1) return Math.floor(interval) + " jours";
+  interval = seconds / 3600;
+  if (interval > 1) return Math.floor(interval) + " heures";
+  interval = seconds / 60;
+  if (interval > 1) return Math.floor(interval) + " minutes";
+  return Math.floor(seconds) + " secondes";
+};
