+++ app/src/utils/string.js
@@ -13,3 +13,19 @@ export const slugify = (string) => {
     .replace(/^-+/, "")
     .replace(/-+$/, "");
 };
+
+export const isValidEmail = (email) => {
+  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
+};
+
+export const hasSpecialChar = (string) => {
+  return /[$&+,:;=?@#|'<>.^*()%!-]/.test(string);
+};
+
+export const hasNumber = (string) => {
+  return /[0-9]/.test(string);
+};
+
+export const hasLetter = (string) => {
+  return /[a-zA-Z]/.test(string);
+};
